requirements.txt
streamlit
pandas
plotly
elasticsearch
python-dotenv
